requirements.txt
# FastAPI Framework
fastapi==0.121.1
uvicorn==0.38.0

# Database
mysql-connector-python==9.5.0
sqlalchemy==2.0.44

# Validation and Data Models
pydantic==2.12.4
pydantic_core==2.41.5
email-validator==2.3.0

# Security and Authentication
passlib==1.7.4
bcrypt==3.2.2
pyjwt==2.10.1

# Environment Variables
python-dotenv==1.2.1

# HTTP Client
httpx>=0.28.0

# Dependencies (auto-installed with above packages)
annotated-types==0.7.0
anyio==4.11.0
cffi==2.0.0
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
greenlet==3.2.4
h11==0.16.0
idna==3.11
pycparser==2.23
sniffio==1.3.1
starlette==0.49.3
typing_extensions==4.15.0
